matched count

+++ cpp/core/common.h
@@ -17,7 +17,7 @@ namespace distblas::core {
 
 int divide_and_round_up(int num, int denom);
 
-vector<int> generate_random_numbers(int lower_bound, int upper_bound, int seed,
+vector<uint64_t> generate_random_numbers(int lower_bound, int upper_bound, int seed,
                                     int ns);
 
 void prefix_sum(vector<int> &values, vector<int> &offsets);
